+++ includes/model/class-application.php
@@ -23,7 +23,7 @@ class Application extends Actor {
 	 *
 	 * @var int
 	 */
-	protected $_id = Actors::APPLICATION_ACTOR_ID; // phpcs:ignore PSR2.Classes.PropertyDeclaration.Underscore
+	protected $_id = Actors::APPLICATION_USER_ID; // phpcs:ignore PSR2.Classes.PropertyDeclaration.Underscore
 
 	/**
 	 * Whether the Application is discoverable.
@@ -224,7 +224,7 @@ class Application extends Actor {
 		return array(
 			'id'           => $this->get_id() . '#main-key',
 			'owner'        => $this->get_id(),
-			'publicKeyPem' => Signature::get_public_key_for( Actors::APPLICATION_ACTOR_ID ),
+			'publicKeyPem' => Signature::get_public_key_for( Actors::APPLICATION_USER_ID ),
 		);
 	}